+++ geotrek/urls.py
@@ -59,6 +59,8 @@ if 'geotrek.feedback' in settings.INSTALLED_APPS:
     urlpatterns.append(path('', include('geotrek.feedback.urls')))
 if 'geotrek.sensitivity' in settings.INSTALLED_APPS:
     urlpatterns.append(path('', include('geotrek.sensitivity.urls')))
+if 'geotrek.flatpages' in settings.INSTALLED_APPS:
+    urlpatterns.append(path('flatpages/', include('geotrek.flatpages.urls')))
 if 'geotrek.api' in settings.INSTALLED_APPS:
     urlpatterns.append(path('', include('geotrek.api.v2.urls')))
     if 'geotrek.flatpages' in settings.INSTALLED_APPS and 'geotrek.trekking' in settings.INSTALLED_APPS and 'geotrek.tourism' in settings.INSTALLED_APPS: